+++ api/golang/core/lib/enclaves/kurtosis_mod.go
@@ -3,7 +3,11 @@ package enclaves
 import (
 	"github.com/go-yaml/yaml"
 	"github.com/kurtosis-tech/stacktrace"
-	"io/ioutil"
+	"os"
+)
+
+const (
+	dependenciesUrl = "https://docs.kurtosis.com/reference/starlark-reference/#dependencies"
 )
 
 type KurtosisMod struct {
@@ -15,8 +19,11 @@ type Module struct {
 }
 
 func parseKurtosisMod(kurtosisModFilepath string) (*KurtosisMod, error) {
-	kurtosisModContents, err := ioutil.ReadFile(kurtosisModFilepath)
+	kurtosisModContents, err := os.ReadFile(kurtosisModFilepath)
 	if err != nil {
+		if os.IsExist(err) {
+			return nil, stacktrace.NewError("Couldn't find a '%v' in the root of the module at '%v'. Modules are expected to have a '%v' at root, have a look at '%v' for more", modFilename, kurtosisModFilepath, modFilename, dependenciesUrl)
+		}
 		return nil, stacktrace.Propagate(err, "An error occurred while reading the '%v' file at '%v'", modFilename, kurtosisModFilepath)
 	}